Fix Task Generator
Today I fixed the Task Generator function, which gives us greater flexibility in terms of the kinds of levels we want to give the game.

The GameGenerator still needs some improvements in terms of adapting to what the TaskGenerator provides. For now, we have a game that doesn't have compiler errors :D